+++ packages/fiber/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @react-three/fiber
 
+## 8.11.9
+
+### Patch Changes
+
+- 6b5f572c: fix: treeshake THREE.ColorManagement
+
+## 8.11.8
+
+### Patch Changes
+
+- 350cd3f3: fix(Canvas): inline render-effect
+
+## 8.11.7
+
+### Patch Changes
+
+- 96af62d5: fix: don't overwrite public cameras
+
 ## 8.11.6
 
 ### Patch Changes